refactor: tier 1 quick wins (file splits, wrapper cleanup)
Mechanical, behavior-preserving refactor across three areas:

- config/config.go (1697 → 740): split into peer files in the same
  package — cache.go (cache types + ValidateCacheConfig + env appliers),
  budget.go (BudgetsConfig + parsers + validation), fallback.go
  (FallbackMode + loadFallbackConfig), env.go (reflect-based env
  override walker + expandString/parseBool helpers).
- internal/providers/anthropic/anthropic.go (1759 → 1582): moved 17
  pure JSON DTO declarations to a new types.go.
- internal/admin/handler.go (2439 → 2400): inlined 5 *UnavailableError
  method wrappers at 17 call sites; collapsed 4 identical-shape
  *WriteError wrappers behind a validationWriter factory
  (modelOverrideWriteError kept standalone for its 502 fallback).

Public API unchanged. Full test suite passes (go test ./...).
