+++ config/config.go
@@ -2,24 +2,16 @@
 package config
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io"
-	"log/slog"
-	"math"
 	"os"
 	"path"
-	"reflect"
 	"regexp"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
 
 	"gopkg.in/yaml.v3"
 
-	"gomodel/internal/core"
 	"gomodel/internal/storage"
 )
 
@@ -95,45 +87,6 @@ type RawRetryConfig struct {
 	JitterFactor   *float64       `yaml:"jitter_factor"`
 }
 
-// FallbackMode controls how alternate models are selected when the primary
-// model is unavailable.
-type FallbackMode string
-
-const (
-	FallbackModeOff    FallbackMode = "off"
-	FallbackModeManual FallbackMode = "manual"
-	FallbackModeAuto   FallbackMode = "auto"
-)
-
-// Valid reports whether mode is one of the supported fallback modes.
-func (m FallbackMode) Valid() bool {
-	switch normalizeFallbackMode(m) {
-	case FallbackModeOff, FallbackModeManual, FallbackModeAuto:
-		return true
-	default:
-		return false
-	}
-}
-
-func normalizeFallbackMode(mode FallbackMode) FallbackMode {
-	return FallbackMode(strings.ToLower(strings.TrimSpace(string(mode))))
-}
-
-// ResolveFallbackDefaultMode canonicalizes the global fallback default mode and
-// applies the process default when unset.
-func ResolveFallbackDefaultMode(mode FallbackMode) FallbackMode {
-	mode = normalizeFallbackMode(mode)
-	if mode == "" {
-		return FallbackModeManual
-	}
-	return mode
-}
-
-// FallbackModelOverride holds per-model mode overrides.
-type FallbackModelOverride struct {
-	Mode FallbackMode `yaml:"mode" json:"mode"`
-}
-
 // ModelsConfig holds global model access defaults.
 type ModelsConfig struct {
 	// EnabledByDefault controls whether provider models are available
@@ -188,24 +141,6 @@ func ResolveConfiguredProviderModelsMode(mode ConfiguredProviderModelsMode) Conf
 		return ConfiguredProviderModelsModeFallback
 	}
 	return mode
-}
-
-// FallbackConfig holds translated-route model fallback policy.
-type FallbackConfig struct {
-	// DefaultMode controls the fallback behavior when no per-model override exists.
-	// Supported values: "auto", "manual", "off". Default: "manual".
-	DefaultMode FallbackMode `yaml:"default_mode" env:"FEATURE_FALLBACK_MODE"`
-
-	// ManualRulesPath points to a JSON file that maps source model selectors to
-	// ordered fallback model selector lists. Empty disables manual rules.
-	ManualRulesPath string `yaml:"manual_rules_path" env:"FALLBACK_MANUAL_RULES_PATH"`
-
-	// Overrides controls per-model mode overrides. Keys may be bare models
-	// ("gpt-4o") or provider-qualified public selectors ("azure/gpt-4o").
-	Overrides map[string]FallbackModelOverride `yaml:"overrides"`
-
-	// Manual holds the parsed manual fallback lists loaded from ManualRulesPath.
-	Manual map[string][]string `yaml:"-"`
 }
 
 // AdminConfig holds configuration for the admin API and dashboard UI.
@@ -383,37 +318,6 @@ type UsageConfig struct {
 	RetentionDays int `yaml:"retention_days" env:"USAGE_RETENTION_DAYS"`
 }
 
-// BudgetsConfig holds per-user-path spend limits.
-type BudgetsConfig struct {
-	// Enabled controls whether budget checks are active.
-	// Default: true. Requires usage tracking because spend limits are evaluated
-	// from usage cost records.
-	Enabled bool `yaml:"enabled" env:"BUDGETS_ENABLED"`
-
-	// UserPaths declares budget limits by tracked user path.
-	UserPaths []BudgetUserPathConfig `yaml:"user_paths"`
-}
-
-// BudgetUserPathConfig declares one or more budget limits for a user path.
-type BudgetUserPathConfig struct {
-	Path   string              `yaml:"path"`
-	Limits []BudgetLimitConfig `yaml:"limits"`
-}
-
-// BudgetLimitConfig declares one spend limit for a reset period.
-type BudgetLimitConfig struct {
-	// Period accepts hourly, daily, weekly, or monthly. The resolved period is
-	// persisted as PeriodSeconds in the database.
-	Period string `yaml:"period"`
-
-	// PeriodSeconds can be set directly instead of Period. Standard values are
-	// 3600, 86400, 604800, and 2592000.
-	PeriodSeconds int64 `yaml:"period_seconds"`
-
-	// Amount is the maximum allowed tracked provider spend for the period.
-	Amount float64 `yaml:"amount"`
-}
-
 // StorageConfig holds database storage configuration (used by audit logging, usage tracking, future IAM, etc.)
 type StorageConfig struct {
 	// Type specifies the storage backend: "sqlite" (default), "postgresql", or "mongodb"
@@ -477,401 +381,6 @@ func (c StorageConfig) BackendConfig() storage.Config {
 		cfg.MongoDB.Database = "gomodel"
 	}
 	return cfg
-}
-
-// CacheConfig holds model and response cache configuration.
-type CacheConfig struct {
-	Model    ModelCacheConfig    `yaml:"model"`
-	Response ResponseCacheConfig `yaml:"response"`
-}
-
-// ModelCacheConfig holds cache configuration for model registry.
-// Exactly one of Local or Redis must be non-nil.
-type ModelCacheConfig struct {
-	RefreshInterval int               `yaml:"refresh_interval" env:"CACHE_REFRESH_INTERVAL"`
-	ModelList       ModelListConfig   `yaml:"model_list"`
-	Local           *LocalCacheConfig `yaml:"local"`
-	Redis           *RedisModelConfig `yaml:"redis"`
-}
-
-// LocalCacheConfig holds local file cache configuration.
-type LocalCacheConfig struct {
-	CacheDir string `yaml:"cache_dir" env:"GOMODEL_CACHE_DIR"`
-}
-
-// ModelListConfig holds configuration for fetching the external model metadata registry.
-type ModelListConfig struct {
-	// URL is the HTTP(S) URL to fetch models.json from (empty = disabled)
-	URL string `yaml:"url" env:"MODEL_LIST_URL"`
-}
-
-// RedisModelConfig holds Redis connection configuration for the model registry cache.
-type RedisModelConfig struct {
-	URL string `yaml:"url" env:"REDIS_URL"`
-	Key string `yaml:"key" env:"REDIS_KEY_MODELS"`
-	TTL int    `yaml:"ttl" env:"REDIS_TTL_MODELS"`
-}
-
-// RedisResponseConfig holds Redis connection configuration for the response cache.
-// Uses separate env vars from RedisModelConfig for key and TTL to allow independent
-// configuration. The URL is shared via REDIS_URL to simplify single-Redis deployments;
-// use YAML config if different Redis instances are needed for model and response caches.
-// Env vars are applied in Load via applyResponseSimpleEnv, only when cache.response.simple is present
-// (see RESPONSE_CACHE_SIMPLE_ENABLED for env-only opt-in without YAML).
-type RedisResponseConfig struct {
-	URL string `yaml:"url"`
-	Key string `yaml:"key"`
-	TTL int    `yaml:"ttl"`
-}
-
-// ResponseCacheConfig holds configuration for response cache middleware.
-type ResponseCacheConfig struct {
-	Simple   *SimpleCacheConfig   `yaml:"simple"`
-	Semantic *SemanticCacheConfig `yaml:"semantic"`
-}
-
-// SimpleCacheConfig holds configuration for exact-match response caching.
-// When the simple block is omitted from config.yaml, this layer stays off unless
-// RESPONSE_CACHE_SIMPLE_ENABLED=true is set (e.g. Helm without a response-cache YAML fragment).
-// Omitted enabled (nil) means true whenever the simple block exists.
-type SimpleCacheConfig struct {
-	Enabled *bool                `yaml:"enabled"`
-	Redis   *RedisResponseConfig `yaml:"redis"`
-}
-
-// SemanticCacheConfig holds configuration for the semantic (vector-similarity) response cache.
-// When the semantic block is omitted from config.yaml, this layer stays off unless
-// SEMANTIC_CACHE_ENABLED=true is set. Omitted enabled (nil) means true whenever the semantic block exists.
-// Tuning env vars are applied in Load via applyResponseSemanticEnv when this block exists.
-type SemanticCacheConfig struct {
-	Enabled                 *bool             `yaml:"enabled"`
-	SimilarityThreshold     float64           `yaml:"similarity_threshold"`
-	TTL                     *int              `yaml:"ttl"`
-	MaxConversationMessages *int              `yaml:"max_conversation_messages"`
-	ExcludeSystemPrompt     bool              `yaml:"exclude_system_prompt"`
-	Embedder                EmbedderConfig    `yaml:"embedder"`
-	VectorStore             VectorStoreConfig `yaml:"vector_store"`
-}
-
-// EmbedderConfig selects how embeddings are generated.
-// Provider must match a key in the top-level providers map when semantic
-// caching is active; that provider's api_key and base_url are reused for
-// POST /v1/embeddings. There is no default provider.
-type EmbedderConfig struct {
-	Provider string `yaml:"provider"`
-	Model    string `yaml:"model"`
-}
-
-// VectorStoreConfig selects the vector DB backend.
-// Type must be set when semantic caching is enabled: qdrant, pgvector, pinecone, weaviate.
-type VectorStoreConfig struct {
-	Type     string         `yaml:"type"`
-	Qdrant   QdrantConfig   `yaml:"qdrant"`
-	PGVector PGVectorConfig `yaml:"pgvector"`
-	Pinecone PineconeConfig `yaml:"pinecone"`
-	Weaviate WeaviateConfig `yaml:"weaviate"`
-}
-
-// QdrantConfig holds connection configuration for the Qdrant vector store.
-type QdrantConfig struct {
-	URL        string `yaml:"url"`
-	Collection string `yaml:"collection"`
-	APIKey     string `yaml:"api_key"`
-}
-
-// PGVectorConfig holds connection configuration for the pgvector vector store.
-type PGVectorConfig struct {
-	URL       string `yaml:"url"`
-	Table     string `yaml:"table"`
-	Dimension int    `yaml:"dimension"`
-}
-
-// PineconeConfig holds connection configuration for Pinecone (data-plane HTTP API).
-type PineconeConfig struct {
-	Host      string `yaml:"host"`
-	APIKey    string `yaml:"api_key"`
-	Namespace string `yaml:"namespace"`
-	Dimension int    `yaml:"dimension"`
-}
-
-// WeaviateConfig holds connection configuration for Weaviate.
-type WeaviateConfig struct {
-	URL    string `yaml:"url"`
-	Class  string `yaml:"class"`
-	APIKey string `yaml:"api_key"`
-}
-
-// ValidateCacheConfig validates the cache configuration in c.
-// For the model cache, exactly one backend (Local or Redis) must be configured;
-// having both or neither is an error. When Redis is selected, its URL must be
-// non-empty. Returns a descriptive error if any constraint is violated, or nil
-// if the configuration is valid.
-func ValidateCacheConfig(c *CacheConfig) error {
-	if c == nil {
-		return fmt.Errorf("cache: configuration is required")
-	}
-	m := &c.Model
-	hasLocal := m.Local != nil
-	hasRedis := m.Redis != nil
-
-	if hasLocal && hasRedis {
-		return fmt.Errorf("cache.model: cannot have both local and redis configured; choose one")
-	}
-	if !hasLocal && !hasRedis {
-		return fmt.Errorf("cache.model: must have either local or redis configured")
-	}
-	if hasRedis && m.Redis.URL == "" {
-		return fmt.Errorf("cache.model.redis: URL is required when using redis")
-	}
-
-	sem := c.Response.Semantic
-	if sem != nil && SemanticCacheActive(sem) {
-		vsType := strings.TrimSpace(sem.VectorStore.Type)
-		if vsType == "" {
-			return fmt.Errorf("cache.response.semantic.vector_store.type: required when semantic cache is enabled; use qdrant, pgvector, pinecone, or weaviate")
-		}
-		switch vsType {
-		case "qdrant", "pgvector", "pinecone", "weaviate":
-		default:
-			return fmt.Errorf("cache.response.semantic.vector_store.type: must be one of qdrant, pgvector, pinecone, weaviate; got %q", sem.VectorStore.Type)
-		}
-		if vsType == "qdrant" {
-			if strings.TrimSpace(sem.VectorStore.Qdrant.URL) == "" {
-				return fmt.Errorf("cache.response.semantic.vector_store.qdrant.url: required when using qdrant")
-			}
-			if strings.TrimSpace(sem.VectorStore.Qdrant.Collection) == "" {
-				return fmt.Errorf("cache.response.semantic.vector_store.qdrant.collection: required when using qdrant")
-			}
-		}
-		if vsType == "pgvector" {
-			if strings.TrimSpace(sem.VectorStore.PGVector.URL) == "" {
-				return fmt.Errorf("cache.response.semantic.vector_store.pgvector.url: required when using pgvector")
-			}
-			if sem.VectorStore.PGVector.Dimension <= 0 {
-				return fmt.Errorf("cache.response.semantic.vector_store.pgvector.dimension: must be > 0 when using pgvector")
-			}
-		}
-		if vsType == "pinecone" {
-			if strings.TrimSpace(sem.VectorStore.Pinecone.Host) == "" {
-				return fmt.Errorf("cache.response.semantic.vector_store.pinecone.host: required when using pinecone")
-			}
-			if strings.TrimSpace(sem.VectorStore.Pinecone.APIKey) == "" {
-				return fmt.Errorf("cache.response.semantic.vector_store.pinecone.api_key: required when using pinecone")
-			}
-			if sem.VectorStore.Pinecone.Dimension <= 0 {
-				return fmt.Errorf("cache.response.semantic.vector_store.pinecone.dimension: must be > 0 when using pinecone")
-			}
-		}
-		if vsType == "weaviate" {
-			if strings.TrimSpace(sem.VectorStore.Weaviate.URL) == "" {
-				return fmt.Errorf("cache.response.semantic.vector_store.weaviate.url: required when using weaviate")
-			}
-			if strings.TrimSpace(sem.VectorStore.Weaviate.Class) == "" {
-				return fmt.Errorf("cache.response.semantic.vector_store.weaviate.class: required when using weaviate")
-			}
-		}
-		st := sem.SimilarityThreshold
-		if math.IsNaN(st) || math.IsInf(st, 0) || st <= 0 || st > 1 {
-			return fmt.Errorf("cache.response.semantic.similarity_threshold: must be greater than 0 and at most 1 (yaml: similarity_threshold, env: SEMANTIC_CACHE_THRESHOLD); got %v", st)
-		}
-		if sem.TTL != nil && *sem.TTL < 0 {
-			return fmt.Errorf("cache.response.semantic.ttl: must be >= 0 (yaml: ttl, env: SEMANTIC_CACHE_TTL); got %d", *sem.TTL)
-		}
-		ep := strings.TrimSpace(sem.Embedder.Provider)
-		if ep == "" {
-			return fmt.Errorf("cache.response.semantic.embedder.provider: required when semantic cache is enabled; use a key from the top-level providers map (e.g. openai, gemini)")
-		}
-		if strings.EqualFold(ep, "local") {
-			return fmt.Errorf("cache.response.semantic.embedder.provider: local embedding is not supported; use a named API provider")
-		}
-	}
-	return nil
-}
-
-// SimpleCacheEnabled reports whether the exact-match response cache layer is
-// allowed to run for a non-nil simple config. Omitted enabled means true.
-func SimpleCacheEnabled(s *SimpleCacheConfig) bool {
-	if s == nil {
-		return false
-	}
-	if s.Enabled != nil && !*s.Enabled {
-		return false
-	}
-	return true
-}
-
-// SemanticCacheActive reports whether the semantic response cache should be
-// validated and constructed. The semantic block must be present (YAML or
-// SEMANTIC_CACHE_ENABLED=true); omitted enabled means true.
-func SemanticCacheActive(sem *SemanticCacheConfig) bool {
-	if sem == nil {
-		return false
-	}
-	if sem.Enabled != nil && !*sem.Enabled {
-		return false
-	}
-	return true
-}
-
-func mergeSemanticResponseDefaults(sem *SemanticCacheConfig) {
-	if sem == nil {
-		return
-	}
-	if sem.SimilarityThreshold == 0 {
-		sem.SimilarityThreshold = 0.92
-	}
-	if sem.TTL == nil {
-		sem.TTL = intPtr(3600)
-	}
-	if sem.MaxConversationMessages == nil {
-		sem.MaxConversationMessages = intPtr(3)
-	}
-}
-
-func intPtr(v int) *int { return &v }
-
-func applyResponseSimpleEnv(resp *ResponseCacheConfig) error {
-	v, ok := os.LookupEnv("RESPONSE_CACHE_SIMPLE_ENABLED")
-	if ok && !parseBool(v) {
-		resp.Simple = nil
-		return nil
-	}
-	if resp.Simple == nil {
-		if ok && parseBool(v) {
-			resp.Simple = &SimpleCacheConfig{}
-		} else {
-			return nil
-		}
-	}
-	simple := resp.Simple
-	if ok {
-		b := parseBool(v)
-		simple.Enabled = &b
-	}
-	if u := os.Getenv("REDIS_URL"); u != "" {
-		if simple.Redis == nil {
-			simple.Redis = &RedisResponseConfig{}
-		}
-		simple.Redis.URL = u
-	}
-	if k := os.Getenv("REDIS_KEY_RESPONSES"); k != "" {
-		if simple.Redis == nil {
-			simple.Redis = &RedisResponseConfig{}
-		}
-		simple.Redis.Key = k
-	}
-	if ts := os.Getenv("REDIS_TTL_RESPONSES"); ts != "" {
-		if simple.Redis == nil {
-			simple.Redis = &RedisResponseConfig{}
-		}
-		n, err := strconv.Atoi(ts)
-		if err != nil {
-			return fmt.Errorf("invalid value for REDIS_TTL_RESPONSES: %q is not a valid integer", ts)
-		}
-		simple.Redis.TTL = n
-	}
-	return nil
-}
-
-func applyResponseSemanticEnv(resp *ResponseCacheConfig) error {
-	v, enabledKeySet := os.LookupEnv("SEMANTIC_CACHE_ENABLED")
-	if enabledKeySet && !parseBool(v) {
-		resp.Semantic = nil
-		return nil
-	}
-	if resp.Semantic == nil {
-		if enabledKeySet && parseBool(v) {
-			resp.Semantic = &SemanticCacheConfig{}
-		} else {
-			return nil
-		}
-	}
-	sem := resp.Semantic
-	if enabledKeySet {
-		b := parseBool(v)
-		sem.Enabled = &b
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_THRESHOLD"); val != "" {
-		f, err := strconv.ParseFloat(val, 64)
-		if err != nil {
-			return fmt.Errorf("invalid value for SEMANTIC_CACHE_THRESHOLD: %q is not a valid float", val)
-		}
-		sem.SimilarityThreshold = f
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_TTL"); val != "" {
-		i, err := strconv.Atoi(val)
-		if err != nil {
-			return fmt.Errorf("invalid value for SEMANTIC_CACHE_TTL: %q is not a valid integer", val)
-		}
-		sem.TTL = &i
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_MAX_CONV_MESSAGES"); val != "" {
-		i, err := strconv.Atoi(val)
-		if err != nil {
-			return fmt.Errorf("invalid value for SEMANTIC_CACHE_MAX_CONV_MESSAGES: %q is not a valid integer", val)
-		}
-		sem.MaxConversationMessages = &i
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_EXCLUDE_SYSTEM_PROMPT"); val != "" {
-		sem.ExcludeSystemPrompt = parseBool(val)
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_EMBEDDER_PROVIDER"); val != "" {
-		sem.Embedder.Provider = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_EMBEDDER_MODEL"); val != "" {
-		sem.Embedder.Model = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_VECTOR_STORE_TYPE"); val != "" {
-		sem.VectorStore.Type = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_QDRANT_URL"); val != "" {
-		sem.VectorStore.Qdrant.URL = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_QDRANT_COLLECTION"); val != "" {
-		sem.VectorStore.Qdrant.Collection = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_QDRANT_API_KEY"); val != "" {
-		sem.VectorStore.Qdrant.APIKey = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_PGVECTOR_URL"); val != "" {
-		sem.VectorStore.PGVector.URL = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_PGVECTOR_TABLE"); val != "" {
-		sem.VectorStore.PGVector.Table = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_PGVECTOR_DIMENSION"); val != "" {
-		n, err := strconv.Atoi(val)
-		if err != nil {
-			return fmt.Errorf("invalid value for SEMANTIC_CACHE_PGVECTOR_DIMENSION: %q is not a valid integer", val)
-		}
-		sem.VectorStore.PGVector.Dimension = n
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_PINECONE_HOST"); val != "" {
-		sem.VectorStore.Pinecone.Host = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_PINECONE_API_KEY"); val != "" {
-		sem.VectorStore.Pinecone.APIKey = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_PINECONE_NAMESPACE"); val != "" {
-		sem.VectorStore.Pinecone.Namespace = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_PINECONE_DIMENSION"); val != "" {
-		n, err := strconv.Atoi(val)
-		if err != nil {
-			return fmt.Errorf("invalid value for SEMANTIC_CACHE_PINECONE_DIMENSION: %q is not a valid integer", val)
-		}
-		sem.VectorStore.Pinecone.Dimension = n
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_WEAVIATE_URL"); val != "" {
-		sem.VectorStore.Weaviate.URL = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_WEAVIATE_CLASS"); val != "" {
-		sem.VectorStore.Weaviate.Class = val
-	}
-	if val := os.Getenv("SEMANTIC_CACHE_WEAVIATE_API_KEY"); val != "" {
-		sem.VectorStore.Weaviate.APIKey = val
-	}
-	return nil
 }
 
 // ServerConfig holds HTTP server configuration
@@ -1180,474 +689,6 @@ func applyYAML(cfg *Config) (map[string]RawProviderConfig, error) {
 	}
 
 	return rawProviders, nil
-}
-
-func loadFallbackConfig(cfg *FallbackConfig) error {
-	if cfg == nil {
-		return nil
-	}
-
-	cfg.DefaultMode = ResolveFallbackDefaultMode(cfg.DefaultMode)
-	if !cfg.DefaultMode.Valid() {
-		return fmt.Errorf("fallback.default_mode must be one of: auto, manual, off")
-	}
-
-	if len(cfg.Overrides) > 0 {
-		normalized := make(map[string]FallbackModelOverride, len(cfg.Overrides))
-		for key, override := range cfg.Overrides {
-			key = strings.TrimSpace(key)
-			if key == "" {
-				return fmt.Errorf("fallback.overrides: model key cannot be empty")
-			}
-			if _, exists := normalized[key]; exists {
-				return fmt.Errorf("fallback.overrides: duplicate model key after trimming: %q", key)
-			}
-			override.Mode = normalizeFallbackMode(override.Mode)
-			if override.Mode == "" {
-				return fmt.Errorf("fallback.overrides[%q].mode must be one of: auto, manual, off", key)
-			}
-			if !override.Mode.Valid() {
-				return fmt.Errorf("fallback.overrides[%q].mode must be one of: auto, manual, off", key)
-			}
-			normalized[key] = override
-		}
-		cfg.Overrides = normalized
-	}
-
-	path := strings.TrimSpace(cfg.ManualRulesPath)
-	if path == "" {
-		cfg.Manual = nil
-		return nil
-	}
-
-	raw, err := os.ReadFile(path)
-	if err != nil {
-		return fmt.Errorf("fallback.manual_rules_path: failed to read %q: %w", path, err)
-	}
-
-	expanded := expandString(string(raw))
-	decoded := make(map[string][]string)
-	decoder := json.NewDecoder(strings.NewReader(expanded))
-
-	token, err := decoder.Token()
-	if err != nil {
-		return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: %w", path, err)
-	}
-	delim, ok := token.(json.Delim)
-	if !ok || delim != '{' {
-		return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: top-level JSON value must be an object", path)
-	}
-
-	seenKeys := make(map[string]struct{})
-	for decoder.More() {
-		token, err := decoder.Token()
-		if err != nil {
-			return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: %w", path, err)
-		}
-		key, ok := token.(string)
-		if !ok {
-			return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: object key must be a string", path)
-		}
-		if _, exists := seenKeys[key]; exists {
-			return fmt.Errorf("fallback.manual_rules_path: duplicate JSON key %q in %q", key, path)
-		}
-		seenKeys[key] = struct{}{}
-
-		var rawModels json.RawMessage
-		if err := decoder.Decode(&rawModels); err != nil {
-			return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: %w", path, err)
-		}
-		if bytes.Equal(bytes.TrimSpace(rawModels), []byte("null")) {
-			return fmt.Errorf("fallback.manual_rules_path: null not allowed for %q in %q", key, path)
-		}
-		var models []string
-		if err := json.Unmarshal(rawModels, &models); err != nil {
-			return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: %w", path, err)
-		}
-		decoded[key] = models
-	}
-
-	token, err = decoder.Token()
-	if err != nil {
-		return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: %w", path, err)
-	}
-	delim, ok = token.(json.Delim)
-	if !ok || delim != '}' {
-		return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: top-level JSON value must be an object", path)
-	}
-
-	var trailing json.RawMessage
-	if err := decoder.Decode(&trailing); err != io.EOF {
-		if err != nil {
-			return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: %w", path, err)
-		}
-		return fmt.Errorf("fallback.manual_rules_path: failed to parse %q: unexpected trailing JSON content", path)
-	}
-
-	manual := make(map[string][]string, len(decoded))
-	for key, models := range decoded {
-		key = strings.TrimSpace(key)
-		if key == "" {
-			return fmt.Errorf("fallback.manual_rules_path: model key cannot be empty")
-		}
-		if _, exists := manual[key]; exists {
-			return fmt.Errorf("fallback.manual_rules_path: duplicate manual rule key after trimming: %q", key)
-		}
-		normalized := make([]string, 0, len(models))
-		for _, model := range models {
-			model = strings.TrimSpace(model)
-			if model == "" {
-				continue
-			}
-			normalized = append(normalized, model)
-		}
-		manual[key] = normalized
-	}
-	cfg.Manual = manual
-	return nil
-}
-
-func applyBudgetEnv(cfg *Config) error {
-	if cfg == nil {
-		return nil
-	}
-	if !cfg.Budgets.Enabled {
-		return nil
-	}
-
-	const prefix = "SET_BUDGET_"
-	for _, item := range os.Environ() {
-		key, value, ok := strings.Cut(item, "=")
-		if !ok || !strings.HasPrefix(key, prefix) || strings.TrimSpace(value) == "" {
-			continue
-		}
-		path := budgetEnvPath(key[len(prefix):])
-		limits, err := parseBudgetEnvLimits(value)
-		if err != nil {
-			return fmt.Errorf("invalid value for %s: %w", key, err)
-		}
-		if len(limits) == 0 {
-			continue
-		}
-		entry := BudgetUserPathConfig{
-			Path:   path,
-			Limits: limits,
-		}
-		replaced := cfg.Budgets.UserPaths[:0]
-		for _, existing := range cfg.Budgets.UserPaths {
-			if existing.Path != entry.Path {
-				replaced = append(replaced, existing)
-			}
-		}
-		cfg.Budgets.UserPaths = append(replaced, entry)
-	}
-	return nil
-}
-
-func budgetEnvPath(suffix string) string {
-	suffix = strings.ToLower(strings.TrimSpace(suffix))
-	if suffix == "" {
-		return "/"
-	}
-	segments := make([]string, 0)
-	for _, part := range strings.Split(suffix, "__") {
-		part = strings.TrimSpace(part)
-		if part != "" {
-			segments = append(segments, part)
-		}
-	}
-	if len(segments) == 0 {
-		return "/"
-	}
-	return "/" + strings.Join(segments, "/")
-}
-
-func parseBudgetEnvLimits(raw string) ([]BudgetLimitConfig, error) {
-	raw = strings.TrimSpace(raw)
-	if raw == "" {
-		return nil, nil
-	}
-	if strings.HasPrefix(raw, "{") {
-		values := map[string]float64{}
-		if err := json.Unmarshal([]byte(raw), &values); err != nil {
-			return nil, err
-		}
-		limits := make([]BudgetLimitConfig, 0, len(values))
-		periods := make([]string, 0, len(values))
-		for period := range values {
-			periods = append(periods, period)
-		}
-		sort.Strings(periods)
-		for _, period := range periods {
-			limits = append(limits, BudgetLimitConfig{Period: period, Amount: values[period]})
-		}
-		return limits, nil
-	}
-	if strings.HasPrefix(raw, "[") {
-		var limits []BudgetLimitConfig
-		if err := json.Unmarshal([]byte(raw), &limits); err != nil {
-			return nil, err
-		}
-		return limits, nil
-	}
-
-	fields := strings.FieldsFunc(raw, func(r rune) bool {
-		return r == ',' || r == ';' || r == '\n'
-	})
-	limits := make([]BudgetLimitConfig, 0, len(fields))
-	for _, field := range fields {
-		field = strings.TrimSpace(field)
-		if field == "" {
-			continue
-		}
-		period, amountText, ok := strings.Cut(field, "=")
-		if !ok {
-			period, amountText, ok = strings.Cut(field, ":")
-		}
-		if !ok {
-			return nil, fmt.Errorf("budget limit %q must use period=amount", field)
-		}
-		amount, err := strconv.ParseFloat(strings.TrimSpace(amountText), 64)
-		if err != nil {
-			return nil, fmt.Errorf("budget amount %q is not a valid number", amountText)
-		}
-		limits = append(limits, BudgetLimitConfig{
-			Period: strings.TrimSpace(period),
-			Amount: amount,
-		})
-	}
-	return limits, nil
-}
-
-func validateBudgetConfig(cfg *BudgetsConfig) error {
-	if cfg == nil {
-		return nil
-	}
-	if !cfg.Enabled {
-		return nil
-	}
-	seen := make(map[string]struct{})
-	for pathIdx, entry := range cfg.UserPaths {
-		if strings.TrimSpace(entry.Path) == "" {
-			return fmt.Errorf("budgets.user_paths[%d].path is required", pathIdx)
-		}
-		normalizedPath, err := core.NormalizeUserPath(entry.Path)
-		if err != nil {
-			return fmt.Errorf("budgets.user_paths[%d].path is invalid: %w", pathIdx, err)
-		}
-		if normalizedPath == "" {
-			return fmt.Errorf("budgets.user_paths[%d].path is required", pathIdx)
-		}
-		cfg.UserPaths[pathIdx].Path = normalizedPath
-		for limitIdx, limit := range entry.Limits {
-			if math.IsNaN(limit.Amount) || math.IsInf(limit.Amount, 0) || limit.Amount <= 0 {
-				return fmt.Errorf("budgets.user_paths[%d].limits[%d].amount must be a finite number greater than 0", pathIdx, limitIdx)
-			}
-			seconds := limit.PeriodSeconds
-			if limit.PeriodSeconds <= 0 {
-				parsed, ok := budgetPeriodSeconds(limit.Period)
-				if !ok {
-					return fmt.Errorf("budgets.user_paths[%d].limits[%d].period must be one of hourly, daily, weekly, monthly or period_seconds must be set", pathIdx, limitIdx)
-				}
-				seconds = parsed
-				cfg.UserPaths[pathIdx].Limits[limitIdx].PeriodSeconds = seconds
-			}
-			key := normalizedPath + ":" + strconv.FormatInt(seconds, 10)
-			if _, ok := seen[key]; ok {
-				return fmt.Errorf("duplicate budget for path %s period %d", normalizedPath, seconds)
-			}
-			seen[key] = struct{}{}
-		}
-	}
-	return nil
-}
-
-func applyBudgetDependencies(cfg *Config) {
-	if cfg == nil || !cfg.Budgets.Enabled || cfg.Usage.Enabled {
-		return
-	}
-	cfg.Budgets.Enabled = false
-	slog.Warn("budget management disabled because usage tracking is disabled",
-		"usage_enabled", false,
-		"budgets_enabled", false,
-		"hint", "enable usage tracking to use budgets, or set BUDGETS_ENABLED=false to silence this warning",
-	)
-}
-
-func budgetPeriodSeconds(period string) (int64, bool) {
-	switch strings.ToLower(strings.TrimSpace(period)) {
-	case "hour", "hourly", "hours":
-		return 3600, true
-	case "day", "daily", "days":
-		return 86400, true
-	case "week", "weekly", "weeks":
-		return 604800, true
-	case "month", "monthly", "months":
-		return 2592000, true
-	default:
-		return 0, false
-	}
-}
-
-// applyEnvOverrides walks cfg's struct fields and applies env var overrides
-// based on `env` struct tags. Maps are skipped.
-func applyEnvOverrides(cfg *Config) error {
-	return applyEnvOverridesValue(reflect.ValueOf(cfg).Elem())
-}
-
-// hasEnvDescendants reports whether t (a struct type) contains any field (at
-// any depth) with a non-empty "env" struct tag. Used to decide whether to
-// allocate a nil pointer-to-struct before recursing into it.
-func hasEnvDescendants(t reflect.Type) bool {
-	if t.Kind() != reflect.Struct {
-		return false
-	}
-	for f := range t.Fields() {
-		if f.Tag.Get("env") != "" {
-			return true
-		}
-		ft := f.Type
-		if ft.Kind() == reflect.Pointer {
-			ft = ft.Elem()
-		}
-		if ft.Kind() == reflect.Struct && hasEnvDescendants(ft) {
-			return true
-		}
-	}
-	return false
-}
-
-func applyEnvOverridesValue(v reflect.Value) error {
-	t := v.Type()
-	for i := 0; i < t.NumField(); i++ {
-		field := t.Field(i)
-		fieldVal := v.Field(i)
-
-		if field.Type.Kind() == reflect.Map {
-			continue
-		}
-		if field.Type.Kind() == reflect.Struct {
-			if err := applyEnvOverridesValue(fieldVal); err != nil {
-				return err
-			}
-			continue
-		}
-		if field.Type.Kind() == reflect.Pointer {
-			elemType := field.Type.Elem()
-			if elemType.Kind() != reflect.Struct {
-				continue
-			}
-			if fieldVal.IsNil() {
-				// Only allocate if the pointed-to struct has env-tagged descendants;
-				// otherwise leave it nil so optional config sections stay absent.
-				if !hasEnvDescendants(elemType) {
-					continue
-				}
-				// Allocate a zero-value struct so env vars can populate its fields.
-				newVal := reflect.New(elemType)
-				if err := applyEnvOverridesValue(newVal.Elem()); err != nil {
-					return err
-				}
-				// Only keep the allocation if at least one field was actually set.
-				if !reflect.DeepEqual(newVal.Elem().Interface(), reflect.Zero(elemType).Interface()) {
-					fieldVal.Set(newVal)
-				}
-			} else {
-				if err := applyEnvOverridesValue(fieldVal.Elem()); err != nil {
-					return err
-				}
-			}
-			continue
-		}
-
-		envKey := field.Tag.Get("env")
-		if envKey == "" {
-			continue
-		}
-		envVal := os.Getenv(envKey)
-		if envVal == "" {
-			continue
-		}
-
-		switch field.Type.Kind() {
-		case reflect.String:
-			fieldVal.SetString(envVal)
-		case reflect.Bool:
-			fieldVal.SetBool(parseBool(envVal))
-		case reflect.Slice:
-			if field.Type.Elem().Kind() != reflect.String {
-				continue
-			}
-			items := strings.Split(envVal, ",")
-			values := make([]string, 0, len(items))
-			for _, item := range items {
-				trimmed := strings.TrimSpace(item)
-				if trimmed == "" {
-					continue
-				}
-				values = append(values, trimmed)
-			}
-			fieldVal.Set(reflect.ValueOf(values))
-		case reflect.Int:
-			n, err := strconv.Atoi(envVal)
-			if err != nil {
-				return fmt.Errorf("invalid value for %s (%s): %q is not a valid integer", field.Name, envKey, envVal)
-			}
-			fieldVal.SetInt(int64(n))
-		case reflect.Int64:
-			if field.Type == reflect.TypeFor[time.Duration]() {
-				// time.Duration is represented as int64; accept Go duration strings (e.g. "1s", "500ms").
-				d, err := time.ParseDuration(envVal)
-				if err != nil {
-					return fmt.Errorf("invalid value for %s (%s): %q is not a valid duration", field.Name, envKey, envVal)
-				}
-				fieldVal.SetInt(int64(d))
-			} else {
-				n, err := strconv.ParseInt(envVal, 10, 64)
-				if err != nil {
-					return fmt.Errorf("invalid value for %s (%s): %q is not a valid integer", field.Name, envKey, envVal)
-				}
-				fieldVal.SetInt(n)
-			}
-		case reflect.Float64:
-			f, err := strconv.ParseFloat(envVal, 64)
-			if err != nil {
-				return fmt.Errorf("invalid value for %s (%s): %q is not a valid float", field.Name, envKey, envVal)
-			}
-			fieldVal.SetFloat(f)
-		}
-	}
-	return nil
-}
-
-// expandString expands environment variable references like ${VAR} or ${VAR:-default} in a string.
-func expandString(s string) string {
-	if s == "" {
-		return s
-	}
-	return os.Expand(s, func(key string) string {
-		varname := key
-		defaultValue := ""
-		hasDefault := false
-		if before, after, ok := strings.Cut(key, ":-"); ok {
-			varname = before
-			defaultValue = after
-			hasDefault = true
-		}
-		value := os.Getenv(varname)
-		if value == "" {
-			if hasDefault {
-				return defaultValue
-			}
-			return "${" + key + "}"
-		}
-		return value
-	})
-}
-
-// parseBool returns true if s is "true" or "1" (case-insensitive).
-func parseBool(s string) bool {
-	return strings.EqualFold(s, "true") || s == "1"
 }
 
 // ValidateBodySizeLimit validates a body size limit string.
