+++ internal/providers/anthropic/types.go
@@ -1,0 +1,180 @@
+package anthropic
+
+import "encoding/json"
+
+// anthropicThinking represents the thinking configuration for Anthropic's extended thinking.
+// For 4.6 models: {type: "adaptive"} (budget_tokens omitted).
+// For older models: {type: "enabled", budget_tokens: N}.
+type anthropicThinking struct {
+	Type         string `json:"type"`
+	BudgetTokens int    `json:"budget_tokens,omitempty"`
+}
+
+// anthropicOutputConfig controls effort level for adaptive thinking on 4.6 models.
+type anthropicOutputConfig struct {
+	Effort string `json:"effort,omitempty"`
+}
+
+// anthropicRequest represents the Anthropic API request format
+type anthropicRequest struct {
+	Model        string                 `json:"model"`
+	Messages     []anthropicMessage     `json:"messages"`
+	Tools        []anthropicTool        `json:"tools,omitempty"`
+	ToolChoice   *anthropicToolChoice   `json:"tool_choice,omitempty"`
+	MaxTokens    int                    `json:"max_tokens"`
+	Temperature  *float64               `json:"temperature,omitempty"`
+	System       any                    `json:"system,omitempty"`
+	Stream       bool                   `json:"stream,omitempty"`
+	Thinking     *anthropicThinking     `json:"thinking,omitempty"`
+	OutputConfig *anthropicOutputConfig `json:"output_config,omitempty"`
+}
+
+type anthropicTool struct {
+	Name        string         `json:"name"`
+	Description string         `json:"description,omitempty"`
+	InputSchema map[string]any `json:"input_schema"`
+}
+
+type anthropicToolChoice struct {
+	Type                   string `json:"type"`
+	Name                   string `json:"name,omitempty"`
+	DisableParallelToolUse *bool  `json:"disable_parallel_tool_use,omitempty"`
+}
+
+// anthropicMessage represents a message in Anthropic format
+type anthropicMessage struct {
+	Role    string `json:"role"`
+	Content any    `json:"content"`
+}
+
+type anthropicContentBlock struct {
+	Type         string                  `json:"type"`
+	Text         string                  `json:"text,omitempty"`
+	ID           string                  `json:"id,omitempty"`
+	Name         string                  `json:"name,omitempty"`
+	Input        any                     `json:"input,omitempty"`
+	ToolUseID    string                  `json:"tool_use_id,omitempty"`
+	Content      any                     `json:"content,omitempty"`
+	IsError      bool                    `json:"is_error,omitempty"`
+	Source       *anthropicContentSource `json:"source,omitempty"`
+	CacheControl json.RawMessage         `json:"cache_control,omitempty"`
+}
+
+type anthropicContentSource struct {
+	Type      string `json:"type"`
+	MediaType string `json:"media_type,omitempty"`
+	Data      string `json:"data,omitempty"`
+	URL       string `json:"url,omitempty"`
+}
+
+// anthropicResponse represents the Anthropic API response format
+type anthropicResponse struct {
+	ID         string             `json:"id"`
+	Type       string             `json:"type"`
+	Role       string             `json:"role"`
+	Content    []anthropicContent `json:"content"`
+	Model      string             `json:"model"`
+	StopReason string             `json:"stop_reason"`
+	Usage      anthropicUsage     `json:"usage"`
+}
+
+// anthropicContent represents content in Anthropic response
+type anthropicContent struct {
+	Type      string          `json:"type"`
+	Text      string          `json:"text,omitempty"`
+	Thinking  string          `json:"thinking,omitempty"`
+	Signature string          `json:"signature,omitempty"`
+	ID        string          `json:"id,omitempty"`
+	Name      string          `json:"name,omitempty"`
+	Input     json.RawMessage `json:"input,omitempty"`
+}
+
+// anthropicUsage represents token usage in Anthropic response
+type anthropicUsage struct {
+	InputTokens              int `json:"input_tokens"`
+	OutputTokens             int `json:"output_tokens"`
+	CacheCreationInputTokens int `json:"cache_creation_input_tokens"`
+	CacheReadInputTokens     int `json:"cache_read_input_tokens"`
+}
+
+// anthropicStreamEvent represents a streaming event from Anthropic
+type anthropicStreamEvent struct {
+	Type         string             `json:"type"`
+	Index        int                `json:"index,omitempty"`
+	Delta        *anthropicDelta    `json:"delta,omitempty"`
+	ContentBlock *anthropicContent  `json:"content_block,omitempty"`
+	Message      *anthropicResponse `json:"message,omitempty"`
+	Usage        *anthropicUsage    `json:"usage,omitempty"`
+}
+
+// anthropicDelta represents a delta in streaming response
+type anthropicDelta struct {
+	Type        string `json:"type"`
+	Text        string `json:"text,omitempty"`
+	Thinking    string `json:"thinking,omitempty"`
+	Signature   string `json:"signature,omitempty"`
+	PartialJSON string `json:"partial_json,omitempty"`
+	StopReason  string `json:"stop_reason,omitempty"`
+}
+
+// anthropicModelInfo represents a model in Anthropic's models API response
+type anthropicModelInfo struct {
+	ID          string `json:"id"`
+	Type        string `json:"type"`
+	CreatedAt   string `json:"created_at"`
+	DisplayName string `json:"display_name"`
+}
+
+// anthropicModelsResponse represents the Anthropic models API response
+type anthropicModelsResponse struct {
+	Data    []anthropicModelInfo `json:"data"`
+	FirstID string               `json:"first_id"`
+	HasMore bool                 `json:"has_more"`
+	LastID  string               `json:"last_id"`
+}
+
+type anthropicBatchCreateRequest struct {
+	Requests []anthropicBatchRequest `json:"requests"`
+}
+
+type anthropicBatchRequest struct {
+	CustomID string           `json:"custom_id"`
+	Params   anthropicRequest `json:"params"`
+}
+
+type anthropicBatchRequestCounts struct {
+	Processing int `json:"processing"`
+	Succeeded  int `json:"succeeded"`
+	Errored    int `json:"errored"`
+	Canceled   int `json:"canceled"`
+	Expired    int `json:"expired"`
+}
+
+type anthropicBatchResponse struct {
+	ID                string                      `json:"id"`
+	Type              string                      `json:"type"`
+	ProcessingStatus  string                      `json:"processing_status"`
+	RequestCounts     anthropicBatchRequestCounts `json:"request_counts"`
+	CreatedAt         string                      `json:"created_at"`
+	EndedAt           string                      `json:"ended_at"`
+	CancelInitiatedAt string                      `json:"cancel_initiated_at"`
+}
+
+type anthropicBatchListResponse struct {
+	Data    []anthropicBatchResponse `json:"data"`
+	FirstID string                   `json:"first_id"`
+	LastID  string                   `json:"last_id"`
+	HasMore bool                     `json:"has_more"`
+}
+
+type anthropicBatchResultLine struct {
+	CustomID string `json:"custom_id"`
+	Result   struct {
+		Type    string          `json:"type"`
+		Message json.RawMessage `json:"message,omitempty"`
+		Error   *struct {
+			Type    string `json:"type"`
+			Message string `json:"message"`
+		} `json:"error,omitempty"`
+	} `json:"result"`
+}
